requirements.txt
django
gunicorn
django-heroku
djangorestframework
channels
django-cors-headers